No longer garbles some unicode text
The main logic in value for the Python 3 case has had to be changed, as composing with bytes removed the error, but garbled 
the output.

This garbling was actually in v, as it expects bytes to be read from a file in 'b' mode.  bytes is str in Python 2, but not in Iron Python 2, 
so the logic in v was only correct for C Python (as it was designed for), not for Iron Python.